+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### [1.7.0]
+- [Issue #233](https://github.com/prb28/vscode-amiga-assembly/issues/233) Build warnings stop the build
 - [Feat #229](https://github.com/prb28/vscode-amiga-assembly/pull/229) *grahambates* BLTCON Helper
 
 ### [1.6.0] - 2022-07-06